Major changes to mini_app example to output more interesting data.
Previously simply output a spherical data set which did not change which made
example visualization code dull when using mini_app as input. Now outputs a
constantly changing (but repetitive) 3D sine wave surface. Many of the example
steered and monitored variables have changed as a result.

Code also shows how to handle the case of a steered variable altering the size
of a data block to be output (not recommended).