Create modularization framework and modularize "Samples" transport code.
This was done to allow for new methods of communication to be added. This patch
adds modularization for the "Samples" transport layer - used for moving data for
online visualization.

A register_module macro is provided to collect information about all of the
different types of module. This information is used to build all the different
bits of each module and create the separate libraries.

Modules can share common code which is linked into the main ReG_Steer library.
The user can select which module the examples are linked with by default during
the build configuration stage. Modularization is selected by default during the
build but can be turned off. If it is turned off then one module is selected to
be built and linked into the main ReG_Steer library.

IO Proxy code has not been modularized yet as this currently only works with
SOAP/WSRF steering. A new proxy independent of SOAP/WSRF needs to be written
after STEERING is modularized.