Use REG_HAS_MSG_NOSIGNAL to correctly handle the use of flags versus signals for send().
The send() method is wrapped as send_no_signal() with the exact same parameters
and return type. If REG_HAS_MSG_NOSIGNAL is defined then the flags are bitwise
OR'd with MSG_NOSIGNAL, otherwise the flags are passed through untouched. In the
latter case SIGPIPE is caught and ignored to produce the same effect.