Use REG_HAS_MSG_DONTWAIT to correctly implement a cross-platform non-blocking recv.
The recv() method is wrapped as recv_non_block with the exact same parameters
and return type. If REG_HAS_MSG_DONTWAIT is defined then the flags are bitwise
OR'd with MSG_DONTWAIT, otherwise the flags are passed through untouched. In the
latter case non-blocking is toggled using calls to fcntl operating on the socket
directly.