Make the finding and linkage of external library dependencies consistent.
Previously, external library dependencies were being collected in two ways:
 * the REG_EXTERNAL_LIBS variable
 * and lots of separate XXX_LIBRARIES variables

REG_EXTERNAL_LIBS was only being used to populate a couple of lines in config
files such as RealityGridConfig.cmake and reg-config and the separate
XXX_LIBRARIES variables were used to link the examples. This lead to a
divergance of which libraries were being used to link verses those that were
being claimed to be needed to be linked!

This patch unifies everything to only use REG_EXTERNAL_LIBS for both linking
and dependency management. At the same time it also ensures that required
dependencies that are implicit on Linux (such as libm and libz) are found
correctly so as to avoid linker errors on other platforms.