Grid Services code removed and gSoap bindings updated to version 2.7.13.
The Grid Services code (SGS, SGSFactory, ServiceGroupRegistration,
RealityGridTree and RealityGridTreeFactory) is no longer used due to the move
to WSRF from OGSI so it has been removed.

The namespaces used for WSRP, WSRL and addressing have also changed and are
updated in this patch.

The version of gSoap used to generate the bindings has been updated to 2.7.13.