Require OpenSSL for WSRF Steering so that we can always use HTTPS or WSSE if required.
OpenSSL is required for HTTPS and/or WSSE. If the user is using HTTP they still
might use WSSE for some low level security so OpenSSL is still required. As a
result there is almost no point in not using OpenSSL.

All WSRF code that tried to (badly) work around a lack of OpenSSL is removed and
OpenSSL is made a requirement.

 * Main CMakeLists.txt now no longer checks for OpenSSL - moved into a
   WSRF-specific file.
 * Minor changes to mini_steerer to reflect changes to the WSRF setup.
 * Removed references to OpenSSL and WSRF, etc in the generated config scripts.