Add a per-module build step so that each module can have custom build steps.
The build steps are put in the CMake/build directory and are driven by
CMake/build/All.cmake which also sets some global module build steps up. This
is all called from the src directory after the targets have been set up so that
things such as source file and target properties can be set.

Code that was trying (unsuccessfully) to set -fPIC on source files within the
ModulesMacro has been moved to All.cmake where it now works as expected.

WSRF is provided as an example module as those source files require WITH_OPENSSL
and WITH_CDATA to be defined while building.