Remove the proxy code for "grid" (WSRF) steering.
This legacy code was either commented out in the source files or was disabled
via the use of Proxy.available == REG_FALSE. This meant that a lot of code was
being compiled that would never be used and may have caused problems in the
long run.

Various struct members also removed as part of this as they were included to
support the proxy in the first place.