Modularize IOProxy Samples Transport code.
A new module, Samples Proxy, is added to allow the transport of IO data via a
proxy. This was originally referred to as IOProxy and resided in
ReG_Steer_Appside_IOProxy.[ch].

Much of the functionality of the Proxy code was provided by various bits of the
Sockets code so to promote code reuse the common code has been moved to a new
file. ReG_Steer_Samples_Transport_Sockets_Shared.c is built into both the Socket
and Proxy Samples libraries as both require it but can not be used together.

All required build infrastructure is also added sush as a dependency file
(CMake/deps/Proxy.cmake) and build file (CMake/build/Proxy.cmake). The build
file ensures the Java IOProxy code is compiled into a jar.

The java code that comprises the actual proxy is also updated to compile cleanly
for Java 6. Most changes were limited to things relating to generics and
autoboxing.