Use REG_HAS_MSG_WAITALL to correctly implement a cross-platform blocking recv.
The recv() method is wrapped as recv_wait_all with exactly the same parameters
and return type. If REG_HAS_MSG_WAITALL is defined then the flags are bitwise
OR'd with MSG_WAITALL, otherwise the flags are passed through untouched. In the
latter case blocking is achieved by using receive loop to collect the required
amount of data from the socket.