Fixed behaviour when REG_STEER_DIRECTORY is not set, or set incorrectly.
If REG_STEER_DIRECTORY is not set a warning message is printed to the effect
that a default of ./ will be used, but the library was exiting anyway.

Also, if REG_STEER_DIRECTORY was set, but to an invalid path, then the same
warning message was printed and the library exited.

These two issues have been fixed as follows: The first now does not cause the
library to exit but the second does, with the default of ./ removed.

Reported here:
http://code.google.com/p/computational-steering/issues/detail?id=2