Move file-based config into file modules and make scratch dir configurable.
Since the move to modularization we should be at pains to make sure that
configuration data intended for each module is not polluting the common
code-base. There was a reference to the file-based steering directory location
in Steerer_connection_table_type even though such references for sockets- and
WSRF-based steering were removed.

This patch removes this anomally and further introduces the idea of a common
scratch directory (that can be used by all modules) that provides the base
for file-based steering messages from now on.

A default scratch dir of ./ is now defined by REG_SCRATCH_DEFAULT which can be
customized during the build via cmake.