Fix invalid XML schema and invalid XML messages.
The XML schema for the messages that the steering library uses was itself
invalid according to xmllint (provided with libXML2 2.6.31).

Also, the messages were not validating against the (fixed) schema due to the use
of the SchemaLocation attribute in the root element. This has now been removed
pending validation by another method - maybe in the steering library code - if
required. A further result of this change is that the code to locate the schema
via REG_STEER_HOME is redundant and so removed.

Additionally, some log messages were being sent with a sequence number in them,
via the <Seq_num> tag, but this tag was not present in the schema. This has now
been added as a choice between <Seq_num> and <key> for messages of the type
LogEntryType.

Finally, parseSteerMessage was searching for the wrong attribute in "msg_UID",
rather than "UID". Fixed.