Streamline XML parsing method signatures.
The methods used to parse bits of XML messages were all passing a namespace
parameter from one to the next when this was not being used at all. There was a
namespace check commented out at the start of the parsing process that would
catch any issues early, before parsing takes place, so this has been reinstated.