Reduce the amount of output from sink.
The 'sink' example application was printing out all the data it received. When
it was connected to mini_app this meant that it was printing out 10000 floats
each time which was hiding any useful debugging output in the terminal.

Now it simply prints a message with the type of the data and the number of
items.