Port sockets code to native Windows.
Add Winsock library startup and shutdown functions to both Steering and Samples
transports. Use ioctlsocket() to provide functionality provided by fnctl() on
POSIX OS's. Ensure that signals code (for SIGPIPE) is not used as Windows does
not know about it. Use closesocket() in all code and alias this to close for
other OS's.