Complete the generation of LaTeX documentation.
Doxygen generates the required inputs to be built by LaTeX but doesn't
actually run LaTeX itself. It does create a Makefile to do this though so
this commit adds a command to run it and actually generate the final PDF
file. As a result the install step has also been fixed to just install the
PDF file and not all the individual sources as was happening before.