Create a new, separate, config header for the examples.
This is so that the examples don't need to pull in ReG_Steer_Config.h which is
bad practice.

One configuration item, REG_HAS_GETPASS, is only ever used by the examples so is
moved out of ReG_Steer_Config.h and into the new file, ReG_Examples_Config.h.